Openstack Update and Kubernetes version based templates.
This change updates the openstack provider with multi-master
settings, includes issue fixes for #164. This change also uses
a new template scheme that bases templates off kubernetes version
and not provider. This allows us to have certified kubernetes versions.

NOTE:
In the new unified template scheme, the following options have been moved
out of the provider config and up to the root kube config.

- kube_master_count
- ssh_pub_key

Other objects have been moved but they are readonly.

i.e.

 {
   "cloud_account_name": "openstack",
   "master_node_size": "m1.small",
   "name": "cheese",
   "node_sizes": [
     "m1.small"
   ],
   "openstack_config": {
     "image_name": "CoreOS",
     "region": "myregion",
     "public_gateway_id": "mygateway"
   },
  "ssh_pub_key": "mypubsshkey"
 }
